2 typedefs added, client and employee
