+++ server/schema/resolvers.js
@@ -1,0 +1,10 @@
+const resolvers = {
+  Query: {
+    thoughts: async (parent, { username }) => {
+      const params = username ? { username } : {};
+      return Thought.find(params).sort({ createdAt: -1 });
+    },
+  },
+};
+
+module.exports = resolvers;
