--- server/schemas/resolvers.js
@@ -1,9 +1,0 @@
-const resolvers = {
-  Query: {
-    helloWorld: () => {
-      return "Hello world!";
-    },
-  },
-};
-
-module.exports = resolvers;
